# src/main/resources/openapi/openapi.yml
openapi: 3.0.3
info:
  title: Users Reactive API
  version: 1.0.0
  description: |
    Microservicio reactivo para la gestión de usuarios.
    El contrato OpenAPI es la fuente principal para los endpoints expuestos.

servers:
  - url: /

tags:
  - name: Users
    description: Operaciones de usuarios

paths:
  /api/v1/users:
    get:
      tags:
        - Users
      operationId: listUsers
      summary: Listar usuarios
      description: Retorna todos los usuarios registrados.
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Users
      operationId: createUser
      summary: Crear usuario
      description: Crea un usuario y retorna su información.

      parameters:
        - in: header
          name: X-Request-Id
          schema:
            type: string
          required: false
          description: Identificador idempotente de la solicitud

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'

      responses:
        '201':
          description: Usuario creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '409':
          description: Correo ya registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '415':
          description: Content-Type no soportado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/users/{userId}:
    get:
      tags:
        - Users
      operationId: getUserById
      summary: Obtener usuario por id
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Users
      operationId: updateUser
      summary: Actualizar usuario
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Correo ya registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Users
      operationId: deleteUser
      summary: Eliminar usuario
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Usuario eliminado
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  parameters:
    UserId:
      in: path
      name: userId
      required: true
      description: Identificador UUID del usuario
      schema:
        type: string
        format: uuid

  schemas:

    UserCreateRequest:
      type: object
      required:
        - name
        - email
        - password
        - phones
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Juan Rodriguez"

        email:
          type: string
          format: email
          maxLength: 150
          pattern: '^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,}$'
          example: "juan@rodriguez.org"

        password:
          type: string
          minLength: 8
          maxLength: 100
          pattern: '^(?=.*[A-Z])(?=.*[a-z])(?=.*\d).{8,}$'
          example: "Abc12345"

        phones:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PhoneRequest'

    UserUpdateRequest:
      type: object
      required:
        - name
        - email
        - password
        - phones
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Juan Rodriguez"

        email:
          type: string
          format: email
          maxLength: 150
          pattern: '^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,}$'
          example: "juan@rodriguez.org"

        password:
          type: string
          minLength: 8
          maxLength: 100
          pattern: '^(?=.*[A-Z])(?=.*[a-z])(?=.*\d).{8,}$'
          example: "Abc12345"

        active:
          type: boolean
          default: true

        phones:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PhoneRequest'

    PhoneRequest:
      type: object
      required:
        - number
        - cityCode
        - countryCode
      properties:
        number:
          type: string
          example: "1234567"

        cityCode:
          type: string
          example: "1"

        countryCode:
          type: string
          example: "57"

    UserResponse:
      type: object
      required:
        - id
        - name
        - email
        - active
        - createdAt
        - updatedAt
        - phones
      properties:
        id:
          type: string
          format: uuid

        name:
          type: string

        email:
          type: string
          format: email

        active:
          type: boolean

        createdAt:
          type: string
          format: date-time

        updatedAt:
          type: string
          format: date-time

        phones:
          type: array
          items:
            $ref: '#/components/schemas/PhoneResponse'

    PhoneResponse:
      type: object
      required:
        - number
        - cityCode
        - countryCode
      properties:
        number:
          type: string

        cityCode:
          type: string

        countryCode:
          type: string

    ErrorResponse:
      type: object
      required:
        - message
        - path
        - timestamp
      properties:
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time